+++ CHANGELOG.md
@@ -40,6 +40,10 @@
 - BREAKING: always move cursor to the right of left surrounding in `add()`, `delete()`, and `replace()` (instead of moving only if it was on the same line as left surrounding).
 - Update process of getting user input: allow `<C-c>` to cancel and make empty string a valid input.
 
+## mini.tabline
+
+- Show quickfix/loclist buffers with special `*quickfix*` label.
+
 
 # Version 0.3.0